rename files to TitleCase
