+++ packages/home/src/components/HomePage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import PageWithHeader, { Head as CommonHead } from './shared/page_with_header';
+import PageWithHeader, { Head as CommonHead } from './shared/PageWithHeader';
 import profileImage from '../img/profile.png';
 
 import { styled } from 'styled-components';
