--- packages/home/src/components/apps_page.scss
@@ -1,27 +1,0 @@
-@import '../styles/responsive_utilities.scss';
-
-.apps-page {
-  .content {
-    overflow: hidden;
-    padding-bottom: 2em;
-
-    .tag-row {
-      display: none;
-    }
-  }
-
-  @include whenDesktop {
-    .content {
-      max-width: 1200px;
-      margin: auto;
-
-      .tag-row {
-        display: block;
-        margin: 0 15px 5px 15px;
-        padding: 0;
-        overflow: hidden;
-        white-space: normal;
-      }
-    }
-  }
-}
