--- packages/home/src/components/shared/footer.scss
@@ -1,29 +1,0 @@
-.footer {
-  margin-top: 30px;
-  background-color: rgba(#fff, 0.3);
-  font-size: 16px;
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-
-  .navigation {
-    display: flex;
-
-    .nav-link {
-      color: white;
-      margin: 15px 0px 0px 15px;
-      height: 20px;
-
-      + .nav-link {
-        border-left: 1px solid white;
-        margin-left: 10px;
-        padding-left: 10px;
-      }
-    }
-  }
-
-  .copyright {
-    margin: 15px;
-    text-align: right;
-  }
-}
