--- packages/home/src/components/shared/page_with_header.scss
@@ -1,6 +1,0 @@
-.page-with-header {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 100vh;
-}
