+++ packages/home/src/pages/apps.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/display-name */
 import React from 'react';
 
-import AppsPage, { Head } from '../components/apps_page';
+import AppsPage, { Head } from '../components/AppsPage';
 
 export { Head };
 export default () => <AppsPage />;
