+++ packages/home/src/pages/games/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 import React from 'react';
-import GamesPage, { Head } from '../../components/games_page';
+import GamesPage, { Head } from '../../components/GamesPage';
 
 export { Head };
 export default () => <GamesPage />;
