+++ packages/home/src/pages/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 import React from 'react';
-import HomePage, { Head } from '../components/home_page';
+import HomePage, { Head } from '../components/HomePage';
 
 export { Head };
 export default () => <HomePage />;
